+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GetStakesById builder supporting Sui RPC API `suix_getStakesByIds`
 - MultiSig and Sponsor SuiTransaction test
+- Publish SuiTransaction test
 
 ### Fixed